Consolidate tests in TestBase
